requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
python-dotenv
Werkzeug
Markdown
python-telegram-bot